refactor: update itemapi thing
